# class-backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Class Operations API
  version: 1.5.0
  license:
    name: Proprietary
    identifier: LicenseRef-ClassOps-Proprietary
  description: |
    API cho các màn React hiện có: xác thực, lập lịch lớp và vận hành giáo viên.
    JWT có `accountScope=PLATFORM|TENANT`; tenant chỉ được suy ra cho phiên TENANT.
    API tạo lớp/ca/override không nhận link online theo DEC-046.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /accounts:
    get:
      operationId: listTenantAccounts
      summary: Tìm, lọc và phân trang tài khoản tenant
      tags: [Tenant accounts]
      parameters: [{$ref: "#/components/parameters/Search"}, {$ref: "#/components/parameters/Page"}, {$ref: "#/components/parameters/PageSize"}, {name: status, in: query, schema: {type: string}}, {name: role, in: query, schema: {type: string}}, {name: profileType, in: query, schema: {type: string}}]
      responses: {"200": {description: PageResponse tài khoản trong tenant JWT}}
    post:
      operationId: createTenantAccount
      summary: Tạo tài khoản và hồ sơ tối thiểu
      tags: [Tenant accounts]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/CreateAccount"}}}}
      responses: {"201": {description: Tài khoản READY + mật khẩu 123456}, "409": {$ref: "#/components/responses/Conflict"}}
  /accounts/{id}:
    get:
      operationId: getTenantAccount
      summary: Xem một tài khoản trong phạm vi quyền
      tags: [Tenant accounts]
      parameters: [{$ref: "#/components/parameters/Id"}]
      responses: {"200": {description: Chi tiết tài khoản}, "404": {$ref: "#/components/responses/NotFound"}}
    patch:
      operationId: updateTenantAccount
      summary: Sửa tên, email, vai trò và phụ huynh
      tags: [Tenant accounts]
      parameters: [{$ref: "#/components/parameters/Id"}]
      responses: {"200": {description: Hồ sơ/vai trò đã cập nhật}, "409": {$ref: "#/components/responses/Conflict"}}
  /accounts/{id}/status:
    patch:
      operationId: changeTenantAccountStatus
      summary: Khóa hoặc mở tài khoản tenant
      tags: [Tenant accounts]
      parameters: [{$ref: "#/components/parameters/Id"}]
      requestBody: {$ref: "#/components/requestBodies/StatusChange"}
      responses: {"200": {description: Trạng thái đã cập nhật}, "409": {$ref: "#/components/responses/Conflict"}}
  /accounts/{id}/credential-resets:
    post:
      operationId: resetTenantAccountCredential
      summary: Reset credential và thu hồi JWT cũ
      tags: [Tenant accounts]
      parameters: [{$ref: "#/components/parameters/Id"}]
      requestBody: {$ref: "#/components/requestBodies/CredentialReset"}
      responses: {"200": {description: Mật khẩu 123456 và token cũ bị thu hồi}}
...
  /classes/{classId}:
    parameters:
      - {$ref: "#/components/parameters/ClassId"}
    get:
      operationId: getClass
      tags: [Class scheduling]
      summary: Chi tiết lớp
      responses:
        "200": {description: Chi tiết lớp, content: {application/json: {schema: {$ref: "#/components/schemas/ClassDetail"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
    put:
      operationId: updateClassDraft
      tags: [Class scheduling]
      summary: Cập nhật lớp Draft
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ClassDraftInput"}
      responses:
        "200":
          description: Draft mới
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ClassDraftRecord"}
        "409": {$ref: "#/components/responses/Conflict"}
...
  /classes/{classId}/enrollments:
    parameters:
      - {$ref: "#/components/parameters/ClassId"}
    get:
      operationId: listClassEnrollments
      tags: [Class lifecycle and enrollments]
      summary: Roster hiện tại hoặc lịch sử enrollment
      parameters:
        - {name: scope, in: query, schema: {type: string, enum: [Active, History], default: Active}}
        - {$ref: "#/components/parameters/Search"}
        - {$ref: "#/components/parameters/Page"}
        - {$ref: "#/components/parameters/PageSize"}
      responses:
        "200": {description: Trang enrollment, content: {application/json: {schema: {$ref: "#/components/schemas/EnrollmentPage"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
    post:
      operationId: addClassEnrollments
      tags: [Class lifecycle and enrollments]
      summary: Thêm hàng loạt nguyên tử và tạo học phí gốc
      description: |
        Chỉ áp dụng ở Scheduled/Active/AwaitingClose. Toàn bộ lô rollback nếu một học sinh lỗi.
        Cảnh báo sức chứa hoặc trùng lịch có ID và phải được gửi lại trong acknowledgedWarningIds.
      parameters:
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/AddEnrollmentsInput"}}}}
      responses:
        "201": {description: Enrollment và học phí đã tạo, content: {application/json: {schema: {$ref: "#/components/schemas/AddEnrollmentsResult"}}}}
        "400": {$ref: "#/components/responses/ValidationError"}
        "409": {$ref: "#/components/responses/Conflict"}
  /classes/{classId}/enrollment-candidates:
    get:
      operationId: listEnrollmentCandidates
      tags: [Class lifecycle and enrollments]
      summary: Học sinh tenant chưa có enrollment active trong lớp
      parameters:
        - {$ref: "#/components/parameters/ClassId"}
        - {$ref: "#/components/parameters/Search"}
        - {$ref: "#/components/parameters/Page"}
        - {$ref: "#/components/parameters/PageSize"}
      responses:
        "200": {description: Trang ứng viên, content: {application/json: {schema: {$ref: "#/components/schemas/EnrollmentCandidatePage"}}}}
  /classes/{classId}/enrollments/{enrollmentId}:
    patch:
      operationId: endClassEnrollment
      tags: [Class lifecycle and enrollments]
      summary: Ghi nhận rời lớp hoặc chuyển lớp có hiệu lực ngay
      parameters:
        - {$ref: "#/components/parameters/ClassId"}
        - {name: enrollmentId, in: path, required: true, schema: {type: string, format: uuid}}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/EndEnrollmentInput"}}}}
      responses:
        "200": {description: Enrollment đã kết thúc, content: {application/json: {schema: {$ref: "#/components/schemas/EndEnrollmentResult"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /classes/{classId}/status:
    patch:
      operationId: changeClassLifecycleStatus
      tags: [Class lifecycle and enrollments]
      summary: Đóng, mở lại hoặc hủy lớp
      parameters:
        - {$ref: "#/components/parameters/ClassId"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/ChangeClassStatusInput"}}}}
      responses:
        "200": {description: Trạng thái mới, content: {application/json: {schema: {$ref: "#/components/schemas/ChangeClassStatusResult"}}}}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Conflict"}
  /students/me/classes:
    get:
      operationId: listOwnStudentClasses
      tags: [Student learning]
      summary: Lớp còn quyền và thẻ lịch sử bị khóa của học sinh
      parameters:
        - {name: access, in: query, schema: {type: string, enum: [All, Accessible, Locked], default: All}}
        - {$ref: "#/components/parameters/Page"}
        - {$ref: "#/components/parameters/PageSize"}
      responses:
        "200": {description: Trang lớp của học sinh, content: {application/json: {schema: {$ref: "#/components/schemas/StudentClassPage"}}}}
  /students/me/classes/{classId}:
    get:
      operationId: getOwnStudentClass
      tags: [Student learning]
      summary: Metadata lớp khi học sinh còn quyền
      parameters:
        - {$ref: "#/components/parameters/ClassId"}
      responses:
        "200": {description: Chi tiết lớp, content: {application/json: {schema: {$ref: "#/components/schemas/StudentClassDetail"}}}}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
  /students/me/classes/{classId}/sessions:
    get:
      operationId: listOwnStudentClassSessions
      tags: [Student learning]
      summary: Nội dung, điểm danh, record, nhận xét và điểm kiểm tra của chính học sinh
      parameters:
        - {$ref: "#/components/parameters/ClassId"}
        - {$ref: "#/components/parameters/Page"}
        - {$ref: "#/components/parameters/PageSize"}
      responses:
        "200": {description: Trang buổi học, content: {application/json: {schema: {$ref: "#/components/schemas/StudentSessionPage"}}}}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
...
    ClassId: {name: classId, in: path, required: true, schema: {type: string, format: uuid}}
...
    Search: {name: search, in: query, schema: {type: string, default: ""}}
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    PageSize: {name: pageSize, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 150}
  requestBodies:
    StatusChange:
      required: true
      content: {application/json: {schema: {type: object, required: [status, reason, version], properties: {status: {type: string, enum: [ACTIVE, LOCKED]}, reason: {type: string}, version: {type: integer, minimum: 0}}}}}
    CredentialReset:
      required: true
      content: {application/json: {schema: {type: object, required: [reason, version], properties: {reason: {type: string}, version: {type: integer, minimum: 0}}}}}
  responses:
    ValidationError:
      description: Dữ liệu không hợp lệ
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
...
    Forbidden:
      description: Không có quyền
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    NotFound:
      description: Không tìm thấy tài nguyên trong tenant
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    Conflict:
      description: |
        Preview stale, TEACHER_OVERLAP, ROOM_OVERLAP, STUDENT_OVERLAP,
        optimistic lock hoặc idempotency conflict.
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
...
  schemas:
...
    CreateAccount:
      type: object
      required: [profileType, username, displayName, roles]
      properties:
        profileType: {type: string, enum: [STAFF, TEACHER, STUDENT]}
        username: {type: string}
        displayName: {type: string}
        email: {type: string, format: email}
        roles: {type: array, minItems: 1, items: {type: string}}
        parentName: {type: string}
        parentPhone: {type: string}
...
    WeeklyPattern:
      type: object
      description: Không có URL online.
      required: [id, weekday, startTime, endTime, mode]
      properties:
        id: {type: string}
        weekday: {type: integer, minimum: 1, maximum: 7, description: ISO Monday=1}
        startTime: {type: string, pattern: '^\d{2}:\d{2}(:\d{2})?$', example: "19:00"}
        endTime: {type: string, pattern: '^\d{2}:\d{2}(:\d{2})?$', example: "20:30"}
        mode: {type: string, enum: [IN_PERSON, ONLINE]}
        roomId: {type: [string, "null"], format: uuid}
    SessionOverride:
      type: object
      description: Chỉ đổi hình thức/phòng của ứng viên; không có URL.
      required: [sessionKey, mode]
      properties:
        sessionKey: {type: string, example: "mon-evening@2026-08-03"}
        mode: {type: string, enum: [IN_PERSON, ONLINE]}
        roomId: {type: [string, "null"], format: uuid}
    ClassDraftInput:
      type: object
      required:
        - name
        - primaryTeacherId
        - startDate
        - totalSessions
        - tuitionAmount
        - hourlyRate
        - defaultMode
      properties:
        name: {type: string, minLength: 1}
        description: {type: string, default: ""}
        primaryTeacherId: {type: string, format: uuid}
        startDate: {type: string, format: date}
        totalSessions: {type: integer, minimum: 1}
        tuitionAmount: {type: number, exclusiveMinimum: 0}
        hourlyRate: {type: number, exclusiveMinimum: 0}
        capacity: {type: [integer, "null"], minimum: 1}
        defaultMode: {type: string, enum: [IN_PERSON, ONLINE]}
        studentIds: {type: array, uniqueItems: true, items: {type: string, format: uuid}}
        patterns: {type: array, items: {$ref: "#/components/schemas/WeeklyPattern"}}
        overrides: {type: array, items: {$ref: "#/components/schemas/SessionOverride"}}
    ClassDraftRecord:
      allOf:
        - {$ref: "#/components/schemas/ClassDraftInput"}
        - type: object
          required: [id, code, status, version]
          properties:
            id: {type: string, format: uuid}
            code: {type: string}
            status: {const: Draft}
            version: {type: integer, format: int64}
...
    AttendanceStatus:
      type: string
      enum: [PRESENT, LATE, LEFT_EARLY, ABSENT_EXCUSED, ABSENT_UNEXCUSED]
...
    LifecycleWarning:
      type: object
      required: [id, code, message, studentIds]
      properties:
        id: {type: string}
        code: {type: string, enum: [CLASS_CAPACITY, ROOM_CAPACITY, STUDENT_SCHEDULE_CONFLICT, MISSING_ATTENDANCE, MISSING_RECORD]}
        message: {type: string}
        studentIds: {type: array, items: {type: string, format: uuid}}
    ClassDetail:
      type: object
      required: [id, code, name, status, version, allowedTransitions, closeReadiness, futureSessionCount]
      properties:
        id: {type: string, format: uuid}
        code: {type: string}
        name: {type: string}
        status: {type: string, enum: [Draft, Scheduled, Active, AwaitingClose, Closed, Cancelled]}
        version: {type: integer, format: int64, minimum: 0}
        allowedTransitions:
          type: array
          items: {type: string, enum: [Closed, AwaitingClose, Cancelled]}
        closeReadiness:
          type: object
          required: [missingAttendanceCount, missingRecordCount, warnings]
          properties:
            missingAttendanceCount: {type: integer, minimum: 0}
            missingRecordCount: {type: integer, minimum: 0}
            warnings: {type: array, items: {$ref: "#/components/schemas/LifecycleWarning"}}
        futureSessionCount: {type: integer, minimum: 0}
      additionalProperties: true
    EnrollmentStudent:
      type: object
      required: [id, code, name]
      properties:
        id: {type: string, format: uuid}
        code: {type: string}
        name: {type: string}
        parentName: {type: [string, "null"]}
        parentPhone: {type: [string, "null"]}
    EnrollmentItem:
      type: object
      required: [id, student, status, effectiveFrom, version]
      properties:
        id: {type: string, format: uuid}
        student: {$ref: "#/components/schemas/EnrollmentStudent"}
        status: {type: string, enum: [Active, Left, Transferred]}
        effectiveFrom: {type: string, format: date, description: "Ngày vào, bao gồm"}
        effectiveTo: {type: [string, "null"], format: date, description: "Ngày rời, không bao gồm"}
        endReason: {type: [string, "null"]}
        tuitionChargeId: {type: [string, "null"], format: uuid}
        originalTuitionAmount: {type: [number, "null"]}
        tuitionStatus: {type: [string, "null"], enum: [UNPAID, PAID, PARTIALLY_REFUNDED, REFUNDED, null]}
        version: {type: integer, format: int64, minimum: 0}
    EnrollmentPage:
      allOf:
        - {$ref: "#/components/schemas/Page"}
        - type: object
          properties:
            items: {type: array, items: {$ref: "#/components/schemas/EnrollmentItem"}}
    EnrollmentCandidate:
      allOf:
        - {$ref: "#/components/schemas/EnrollmentStudent"}
    EnrollmentCandidatePage:
      allOf:
        - {$ref: "#/components/schemas/Page"}
        - type: object
          properties:
            items: {type: array, items: {$ref: "#/components/schemas/EnrollmentCandidate"}}
    AddEnrollmentsInput:
      type: object
      required: [studentIds, classVersion]
      properties:
        studentIds: {type: array, minItems: 1, uniqueItems: true, items: {type: string, format: uuid}}
        classVersion: {type: integer, format: int64, minimum: 0}
        acknowledgedWarningIds: {type: array, uniqueItems: true, items: {type: string}}
    AddEnrollmentsResult:
      type: object
      required: [enrollments, classVersion]
      properties:
        enrollments: {type: array, items: {$ref: "#/components/schemas/EnrollmentItem"}}
        classVersion: {type: integer, format: int64, minimum: 0}
    EndEnrollmentInput:
      type: object
      required: [targetStatus, reason, classVersion, enrollmentVersion]
      properties:
        targetStatus: {type: string, enum: [Left, Transferred]}
        reason: {type: string, minLength: 1}
        classVersion: {type: integer, format: int64, minimum: 0}
        enrollmentVersion: {type: integer, format: int64, minimum: 0}
    EndEnrollmentResult:
      type: object
      required: [enrollment, classVersion]
      properties:
        enrollment: {$ref: "#/components/schemas/EnrollmentItem"}
        classVersion: {type: integer, format: int64, minimum: 0}
    ChangeClassStatusInput:
      type: object
      required: [targetStatus, version]
      properties:
        targetStatus: {type: string, enum: [Closed, AwaitingClose, Cancelled]}
        reason: {type: string, description: Bắt buộc khi mở lại hoặc hủy}
        version: {type: integer, format: int64, minimum: 0}
        acknowledgedWarningIds: {type: array, uniqueItems: true, items: {type: string}}
    ChangeClassStatusResult:
      type: object
      required: [classId, status, version, cancelledFutureSessions, acknowledgedWarnings]
      properties:
        classId: {type: string, format: uuid}
        status: {type: string, enum: [Closed, AwaitingClose, Cancelled]}
        version: {type: integer, format: int64, minimum: 0}
        cancelledFutureSessions: {type: integer, minimum: 0}
        acknowledgedWarnings: {type: array, items: {$ref: "#/components/schemas/LifecycleWarning"}}
    StudentClassItem:
      type: object
      required: [id, code, name, teacherName, scheduleSummary, status, access, effectiveFrom, completedSessions, totalSessions]
      properties:
        id: {type: string, format: uuid}
        code: {type: string}
        name: {type: string}
        teacherName: {type: string}
        scheduleSummary: {type: string}
        status: {type: string, enum: [Scheduled, Active, AwaitingClose, Closed, Cancelled]}
        access: {type: string, enum: [Accessible, Locked]}
        effectiveFrom: {type: string, format: date}
        effectiveTo: {type: [string, "null"], format: date}
        completedSessions: {type: integer, minimum: 0}
        totalSessions: {type: integer, minimum: 1}
        expectedEndDate: {type: [string, "null"], format: date}
    StudentClassPage:
      allOf:
        - {$ref: "#/components/schemas/Page"}
        - type: object
          properties:
            items: {type: array, items: {$ref: "#/components/schemas/StudentClassItem"}}
    StudentClassDetail:
      type: object
      required: [id, code, name, description, teacherName, scheduleSummary, status, completedSessions, totalSessions]
      properties:
        id: {type: string, format: uuid}
        code: {type: string}
        name: {type: string}
        description: {type: string}
        teacherName: {type: string}
        scheduleSummary: {type: string}
        status: {type: string, enum: [Scheduled, Active, AwaitingClose]}
        completedSessions: {type: integer, minimum: 0}
        totalSessions: {type: integer, minimum: 1}
        expectedEndDate: {type: [string, "null"], format: date}
    StudentSessionItem:
      type: object
      required: [id, ordinal, startAt, endAt, status, teacherName, lessonName, lessonContent]
      properties:
        id: {type: string, format: uuid}
        ordinal: {type: integer, minimum: 1}
        startAt: {type: string, format: date-time}
        endAt: {type: string, format: date-time}
        status: {type: string}
        teacherName: {type: string}
        lessonName: {type: string}
        lessonContent: {type: string}
        attendanceStatus: {oneOf: [{$ref: "#/components/schemas/AttendanceStatus"}, {type: "null"}]}
        attendanceNote: {type: [string, "null"]}
        recordUrl: {type: [string, "null"], format: uri}
        comment: {type: [string, "null"]}
        testResult:
          oneOf:
            - {$ref: "#/components/schemas/StudentTestResult"}
            - {type: "null"}
    StudentTestResult:
      type: object
      required: [id, testName, score, maxScore, testDate, comment, testComment]
      properties:
        id: {type: string, format: uuid}
        testName: {type: string}
        score: {type: [number, "null"], minimum: 0}
        maxScore: {type: number, exclusiveMinimum: 0}
        testDate: {type: string, format: date}
        comment: {type: string, description: Nhận xét riêng cho kết quả của học sinh}
        testComment: {type: string, description: Ghi chú chung của bài kiểm tra}
    StudentSessionPage:
      allOf:
        - {$ref: "#/components/schemas/Page"}
        - type: object
          properties:
            items: {type: array, items: {$ref: "#/components/schemas/StudentSessionItem"}}
...
    Page:
      type: object
      required: [items, page, pageSize, totalItems, totalPages]
      properties:
        items: {type: array, items: {}}
        page: {type: integer, minimum: 1}
        pageSize: {type: integer, minimum: 1}
        totalItems: {type: integer, format: int64}
        totalPages: {type: integer}
    Problem:
      type: object
      required: [type, title, status, detail, code, message, timestamp, traceId]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string}
        code: {type: string}
        message: {type: string}
        fieldErrors: {type: object, additionalProperties: {type: string}}
        details: {type: object, additionalProperties: true}
        retryAfterSeconds: {type: integer, minimum: 1}
        timestamp: {type: string, format: date-time}
        traceId: {type: string}